py-impl/raahi-agent/pyproject.toml
[project]
name = "raahi-assistant"
version = "0.1.0"
description = "AI Assistant for Raahi with voice interaction, Typesense search, and Vertex AI"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.109.0",
    "uvicorn[standard]>=0.27.0",
    "google-cloud-aiplatform>=1.38.0",
    "google-cloud-texttospeech>=2.14.0",
    "typesense>=0.21.0",
    "pydantic>=2.5.0",
    "pydantic-settings>=2.1.0",
    "redis>=5.0.0",
    "python-multipart>=0.0.6",
    "httpx>=0.26.0",
    "firebase-admin>=6.4.0",
    "aiohttp>=3.9.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.23.0",
    "ruff>=0.1.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["app", "config"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.pytest.ini_options]
asyncio_mode = "auto"
